Team members do not block players from trying to build limited units/buildings anymore.
This way a team member (e.g. an AI in permanent alliance) cannot stop you from trying to build wonders. One downside of this change is that you can now get failgold for wonders more easily in coordination with your team members.
Doing this for projects will require a bit more work since projects are tied to teams only and not players.