Bugfix and additional restrictions in unit gifting.
Fixed bug in AI evaluation of gifted unit for the purpose of relations bonus.
Player cannot gift combat units to third party which in war with rival with whom the player has unbreakable (temporary or permanent) peace treaty.
Also, the receiving player now must satisfy technology requirements (both for the unit itself and for the prerequisite resources).