Fixed how AI and worker actions treat fallout feature.
Fallout had bNoImprovement set to 1 signaling to AI that no improvements can be built there which didn't make sense since fallout can be cleared just like forest and jungle can. This made AI far less likely to scrub fallout (if at all) than to chop jungle although the features are similar. Improvements can now be built directly on top of fallout and fallout will be cleared in the process just like it is with forest and jungle.