Fixed Bomb Shelter effect for non-combat units.
Bomb Shelter used to reduce the probability of nuke destroying a non-combat unit from around 80% to around 2-3% (these probabilities are not immediately obvious and have to be calculated; exercise is left for the reader) because someone was not careful with probabilities. NUKE_NON_COMBAT_DEATH_THRESHOLD was probably picked so that the probability is close to average nuke damage to combat units (which is 79%).
I changed the chance to destroy a non-combat unit to exactly 79%, and Bomb Shelter halves that so those are now exactly the same as average damage to combat units. NUKE_NON_COMBAT_DEATH_THRESHOLD is now unused.