Reworked how terrain damage affects path costs.
Idea is that damaging terrain doesn't have fixed high costs, but that the cost is calculated depending on how much time is needed to recover.
Also black hole gravity filed and wormholes have fixed high costs so that they are avoided.
My previous code that was supposed to imitate original FFP code was bugged, that is now fixed.
What is stil missing is avoiding paths that will certainly destroy the unit through terrain damage.