Changed some details of how starting locations are picked.
Starting location picking system is now the same in every game mode. After randomly assigning preplaced starting locations, it first picks locations for some number of AIs depending on handicap, then for all humans and then for the rest of AIs. Order doesn't depend on player id numbers. This is done this way because locations picked earlier are usually better.
Starting area picking system is changed so that it doesn't overvalue large, but very bad areas. When picking starting location, locations with very little food (before normalization) will be avoided if possible to avoid starting on the edge of very bad terrain (after normalization).
In advanced start, the possibility of automatically exchanging starting location with a teammember under certain conditions is now removed. (Why was that even there?)
Fixed a bug in location evaluation in case when starting location needs to be picked again in advanced start.