Changed colors in scoreboard and wonder list.
Wonder list now uses player text color instead of primary color.
Teamscore and team id are not in brackets anymore. They are also gray if that option is used for dead civs.
Player names of unknown players are also concealed during advanced start. Name, civilization and color of dead players are not concealed.