Adjusted domination thresholds for permanent alliances.
Minimum land percentage is now 100*(1+n)/(3+n)+1 where n is number of players in the team. This used to be 51 for all teams.
For domination population, in addition to 25% lead required, the population of the second best team is scaled with the standard ratio of (1+0.5*(n-1)) values for both teams (with a minimum of 1), i.e. larger lead is required in the case that the second best team has less players.