Fixed a crash when pillaging is intercepted by sea patrol.
If the selection group survives the battle with sea patrol when attempting to pillage a water resource then the game crashes because of the nullpointer dereferencing exception in the later code. This is because combat clears the mission queue. It seems that the rest of the function here can be skipped safely in this situation since post-combat code deals with everything necessary (setting activity to ACTIVITY_AWAKE and dealing with unit selection for the active player).
This bug was most likely not present in BTS and was introduced by karadoc (or by BBAI) when pillaging order was improved and when the bulk of pillage code was moved upwards in the function.