Improved OOSLogger.
It should now properly log OOS events. It should also work after reloading or restarting the game.
It still doesn't create logs for PitBoss servers, but it does create logs for players logged in a PitBoss server.
It also still doesn't create logs for subsequent OOS events if the game is not reloaded or restarted (client-side).